feat: 영속성 유지를 위한 userDetails dto 설정
